feat: Update Property & Delete Profile Logic (#40)
* feat: imeplemnt my properties endpoint

* feat: add methods for getting a property for host

* fix: updated service methods

* fix: update service call

* fix: missing id

* feat: enhanced delete account logic

* fix: Fix property id PK violations in tests

* hotfix: updated for account attribute

---------

+++ src/main/java/com/lunark/lunark/auth/service/AccountService.java
@@ -1,7 +1,12 @@
 package com.lunark.lunark.auth.service;
 
 import com.lunark.lunark.auth.model.Account;
+import com.lunark.lunark.auth.model.AccountRole;
 import com.lunark.lunark.properties.model.Property;
+import com.lunark.lunark.properties.repostiory.IPropertyRepository;
+import com.lunark.lunark.properties.service.IPropertyService;
+import com.lunark.lunark.reservations.model.Reservation;
+import com.lunark.lunark.reservations.model.ReservationStatus;
 import com.lunark.lunark.reviews.model.Review;
 import com.lunark.lunark.auth.repository.IAccountRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +15,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -17,8 +23,16 @@ public class AccountService implements IAccountService {
     @Autowired
     IAccountRepository accountRepository;
 
+    // @Autowired
+    // ReservationRepository reservationRepository;
+
+    @Autowired
+    IPropertyService propertyService;
+
     @Autowired
     PasswordEncoder passwordEncoder;
+
+
 
     @Override
     public Collection<Account> findAll() {
@@ -63,9 +77,56 @@ public class AccountService implements IAccountService {
         return accountRepository.saveAndFlush(account);
     }
 
-    @Override public void delete(Long id) {
+    @Override
+    public boolean delete(Long id) {
+        Optional<Account> accountToRemove = accountRepository.findById(id);
+        if (accountToRemove.isEmpty()) { return false; }
+        Account account = accountToRemove.get();
+        AccountRole accountRole = account.getRole();
+
+        if (isGuestAccount(accountRole)) {
+            // Uncomment when Reservation Service and List are implemented
+            // handleUserAccountDeletion(id);
+        } else {
+            List<Property> propertiesList = propertyService.findAllPropertiesForHost(account.getId());
+            handleHostAccountDeletion(id, propertiesList);
+        }
         accountRepository.deleteById(id);
+        return true;
     }
+
+    private boolean isGuestAccount(AccountRole accountRole) {
+        return accountRole == AccountRole.GUEST;
+    }
+
+    private void handleUserAccountDeletion(Long userId) {
+        // Uncomment when Reservation Service and List are implemented
+        // List<Reservation> reservationList = reservationRepository.getAllReservationsForUser(userId);
+        // if (noAcceptedReservations(reservationList)) {
+        //
+        //     accountRepository.deleteById(userId);
+        // }
+    }
+
+    private void handleHostAccountDeletion(Long hostId, List<Property> propertiesList) {
+        // Uncomment when Reservation Service and List are implemented
+        // List<Reservation> reservationList = reservationRepository.getAllReservationsForPropertiesList(propertiesList);
+        // if (noAcceptedReservations(reservationList)) {
+        //     for (Property property : propertiesList) {
+        //         propertyService.delete(property.getId());
+        //     }
+        //     accountRepository.deleteById(hostId);
+        // }
+        for (Property property : propertiesList) {
+            propertyService.delete(property.getId());
+        }
+        accountRepository.deleteById(hostId);
+    }
+
+    public static boolean noAcceptedReservations(List<Reservation> reservationList) {
+        return reservationList.stream().noneMatch(reservation -> reservation.getStatus() == ReservationStatus.ACCEPTED);
+    }
+
 
 
     @Override
